pyproject.toml
[tool.poetry]
name = "innoterservice"
version = "0.1.0"
description = ""
authors = ["blessedboy14 <dansof25@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
python-jose = "^3.3.0"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.1"}
mysqlclient = "^2.2.4"
requests = "^2.32.2"
ruff = "^0.4.4"
pytest = "^8.2.1"
pytest-django = "^4.8.0"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"
faker = "^25.2.0"
pytest-cov = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
